requirements-dev.txt
black~=19.10b0
coveralls~=1.7.0
fakeredis==1.6.1
flake8~=3.7.8
freezegun~=0.3.12
lupa==1.10
pre-commit~=1.21.0
pre-commit-hooks~=2.4.0
pytest~=5.1.3
pytest-cov~=2.7.1
pytest-localserver~=0.5.0
pytest-mongodb~=2.2.0
pytest-rerunfailures~=9.1.1
pytest-sanic~=1.0.0
pytest-xdist~=1.29.0
pytype~=2019.7.11
